requirements.txt
Flask
psycopg[binary]
python-dotenv
Flask-WTF
WTForms
gunicorn